+++ bin/livereload.dart
@@ -1,12 +1,10 @@
 import 'dart:async';
 import 'dart:io';
 
-import 'package:logging/logging.dart';
-
 import 'package:livereload/livereload.dart';
 
 Future<Null> main(List<String> args) async {
-  Logger.root.onRecord.listen(stdIOLogListener);
+  logger.onRecord.listen(stdIOLogListener);
 
   final results = liveReloadArgParser.parse(args);
   if (results[CliOption.help] == true) {
@@ -16,12 +14,17 @@ Future<Null> main(List<String> args) async {
 
   final buildRunner = new BuildRunnerServeProcess.fromParsed(results)..start();
 
-  new LiveReloadProxyServer.fromParsed(
-      results,
-      buildRunner,
+  final webSocket =
       new LiveReloadWebSocketServer.fromParsed(results, buildRunner.onBuild)
-        ..serve())
-    ..serve();
+        ..serve();
 
-  exit(await buildRunner.exitCode);
+  final proxy =
+      new LiveReloadProxyServer.fromParsed(results, buildRunner, webSocket)
+        ..serve();
+
+  ProcessSignal.SIGINT.watch().take(1).listen((_) {
+    buildRunner.kill();
+    proxy.forceClose();
+    webSocket.forceClose();
+  });
 }