+++ lib/src/server/signals.dart
@@ -1,4 +1,7 @@
-/// A message send through the WebSocket signaling the browser to reload.
+/// A message sent from the browser signaling disconnection.
+const String disconnectSignal = 'LelouchViBritannia';
+
+/// A message sent through the WebSocket signaling the browser to reload.
 ///
 /// [Amaki Sally](http://www.asianjunkie.com/2017/10/04/an-introduction-to-the-entertaining-world-of-amaki-sally-a-member-of-227/)
 const String reloadSignal = 'AmakiSally';